small CSS fix, reverting to default.
line 62 and 74 : background URL from "taverna.png" (my change) to "tavern.png" (default). I had made the url changes as a test for functionality while altering BG behavior. 

Even if taverna.png is in the CSS, the BG swapping works properly. 

I think I did this because having it set to tavern.png in the initial CSS will remove whatever BG was chosen by the user in a previous session. Could be wrong about that.