Updated Right Nav And Added Click-Away Closing
Rewrote the nav HTML/CSS

.nav --> #right-nav-panel
.nav-toggle <label> --> removed, using #nav-toggle:after for styling instead

The main purpose of this rewrite was to allow for a function such that the nav panel will auto-close when the user clicks away from it. The former structure of using a label for a hidden checkbox made this impossible, as the label produced two clicks when clicked on (one for the label, and one for the box), therefore locking the panel if the click-away feature was turned on.

I will add in a 'pin' toggle soon to allow users to hold the nav panel open if they wish.

- Changed default font size from 1rem to 15px;

- removed unused CSS code for <main>,<article>,<header>, and <footer>, which were not used in the HTML