Formatting: Ignore chat separators with markdown
Some chat separators (or dinkuses) cause markdown to be rendered
on the chat window.

Examples include "###" -> h3 and "---" -> metadata.

This can look jarring to the end user as it can interrupt a pleasant
chat experience. Therefore, it makes sense to ignore these lines with
div tags that indicate to Showdown that this string doesn't need markdown.

Signed-off-by: kingbri <bdashore3@proton.me>