World Info: Allow disabling recursion on entries
Recursive scanning is a very great tool used to create a tree hierarchy
of entries. However, some entries should not be included in recursion
due to possible conflicts and resulting leakage in chats.

Add an individual opt-out toggle to exclude the entry from recursive
scanning if the main option is enabled.

Signed-off-by: kingbri <bdashore3@proton.me>